Clearance authentication
with Devise turned off

generated with
  rails generate clearance:install
  rails db:migrate

note any previous users generated with devise would need to have a
remember token set to be able to login as Clearance users!
  User.find(1).reset_remember_token!

the /session is overridden to allow for json format to get user and
ultimately to sign in as well. based on Clearance::SessionsController
implementation and calling super for the html format. This could be all
implemented as
    if status.success? || current_user
      format.html { redirect_back_or url_after_create }
      format.json { render json: @user || current_user, status: :ok }
    else
      format.html do
        flash.now.notice = status.failure_message
        render template: "sessions/new", status: :unauthorized
    format.json { render json: [errors: status.failure_message], status: :unauthorized }